split CashuPaymentService logic into CashuSwapHandler and CashuMeltHandler
