+++ BTCPayserver.Plugins.Cashu.Tests/Unit/FailedTransactionsPollerTests.cs
@@ -34,7 +34,7 @@ public class FailedTransactionsPollerTests(ITestOutputHelper output)
 
     private BTCPayServer.Plugins.Cashu.Services.FailedTransactionsPoller CreatePoller(
         TestDbFactory db) =>
-        new(db, null!, null!,null!, new XunitLogger<BTCPayServer.Plugins.Cashu.Services.FailedTransactionsPoller>(output))
+        new(db, null!, null!,null!, new XunitLogger<BTCPayServer.Plugins.Cashu.Services.FailedTransactionsPoller>(output), null!, null!)
         {
             PollInterval = TimeSpan.FromDays(1), // disable auto-polling in unit tests
         };
@@ -123,7 +123,7 @@ public class FailedTransactionsPollerTests(ITestOutputHelper output)
         // Use real constructor defaults (not CreatePoller which overrides PollInterval)
         var poller = new BTCPayServer.Plugins.Cashu.Services.FailedTransactionsPoller(
             db, null!, null!,null!,
-            new XunitLogger<BTCPayServer.Plugins.Cashu.Services.FailedTransactionsPoller>(output));
+            new XunitLogger<BTCPayServer.Plugins.Cashu.Services.FailedTransactionsPoller>(output), null!, null!);
 
         Assert.Equal(TimeSpan.FromMinutes(2), poller.PollInterval);
     }
@@ -134,7 +134,7 @@ public class FailedTransactionsPollerTests(ITestOutputHelper output)
         var db = TestDbFactory.Create();
         var poller = new BTCPayServer.Plugins.Cashu.Services.FailedTransactionsPoller(
             db, null!, null!,null!,
-            new XunitLogger<BTCPayServer.Plugins.Cashu.Services.FailedTransactionsPoller>(output));
+            new XunitLogger<BTCPayServer.Plugins.Cashu.Services.FailedTransactionsPoller>(output), null!, null!);
 
         Assert.Equal(50, poller.BatchSize);
     }
@@ -145,7 +145,7 @@ public class FailedTransactionsPollerTests(ITestOutputHelper output)
         var db = TestDbFactory.Create();
         var poller = new BTCPayServer.Plugins.Cashu.Services.FailedTransactionsPoller(
             db, null!, null!,null!,
-            new XunitLogger<BTCPayServer.Plugins.Cashu.Services.FailedTransactionsPoller>(output));
+            new XunitLogger<BTCPayServer.Plugins.Cashu.Services.FailedTransactionsPoller>(output), null!, null!);
 
         Assert.Equal(3, poller.MaxConcurrencyPerMint);
     }
